ap-tracker/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
anthropic>=0.50.0
schedule==1.2.2
